requirements.txt
Django>=3.2.5,<3.3
psycopg2==2.9.5
import-export==0.2.67.dev6
django-import-export #==2.7.1
xlrd2>=1.3.4,<1.4
openpyxl>=3.0.9,<3.1
djangorestframework>=3.13.1,<3.14
django-cors-headers>=3.13.0,<4
djangorestframework-simplejwt==5.2.2
PyJWT==2.3.0
django-filter==23.2